[TMail] refoctor: group all rag indexation related classes in one folder
