+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/rag/configuration/RagConfig.java
@@ -16,7 +16,7 @@
  *  more details.                                                   *
  ********************************************************************/
 
-package com.linagora.tmail.mailet.rag;
+package com.linagora.tmail.rag.configuration;
 
 import java.net.MalformedURLException;
 import java.net.URI;
