+++ tmail-backend/tmail-third-party/ai-bot/src/test/java/com/linagora/tmail/rag/configuration/RagConfigTest.java
@@ -15,7 +15,7 @@
  *  PURPOSE. See the GNU Affero General Public License for          *
  *  more details.                                                   *
  ********************************************************************/
-package com.linagora.tmail.mailet.rag;
+package com.linagora.tmail.rag.configuration;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -27,7 +27,7 @@ import org.apache.commons.configuration2.Configuration;
 import org.apache.commons.configuration2.PropertiesConfiguration;
 import org.junit.jupiter.api.Test;
 
-import com.linagora.tmail.mailet.AIBotConfig;
+import com.linagora.tmail.common.chatlanguagemodel.LlmConfig;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
