+++ tmail-backend/tmail-third-party/ai-bot/src/test/resources/listeners.xml
@@ -21,6 +21,6 @@
 
 <listeners>
     <listener>
-        <class>com.linagora.tmail.mailet.rag.RagListener</class>
+        <class>com.linagora.tmail.rag.listener.RagListener</class>
     </listener>
 </listeners>
